backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
openai>=1.50.0
anthropic>=0.30.0
httpx>=0.24.0
python-multipart==0.0.6
python-dotenv==1.0.0